# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Precifica API - Financial management and recipe costing for small food businesses
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Perfil {
  id: string;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  nome?: string | null;
  /** @nullable */
  created_at?: string | null;
}

export interface Produto {
  id: string;
  nome: string;
  /** @nullable */
  categoria?: string | null;
  preco_venda: number;
  custo_mao_obra: number;
  frete: number;
  imposto_pct: number;
  taxa_cartao_pct: number;
  taxa_app_pct: number;
  comissao_pct: number;
  /** @nullable */
  cmv?: number | null;
  /** @nullable */
  margem_contribuicao?: number | null;
  /** @nullable */
  margem_pct?: number | null;
  /** @nullable */
  created_at?: string | null;
}

export interface ProdutoInput {
  nome: string;
  /** @nullable */
  categoria?: string | null;
  preco_venda: number;
  custo_mao_obra: number;
  frete: number;
  imposto_pct: number;
  taxa_cartao_pct: number;
  taxa_app_pct: number;
  comissao_pct: number;
}

export interface Insumo {
  id: string;
  nome: string;
  unidade: string;
  preco_unitario: number;
  fator_correcao: number;
  /** @nullable */
  created_at?: string | null;
}

export interface InsumoInput {
  nome: string;
  unidade: string;
  preco_unitario: number;
  fator_correcao: number;
}

export interface FichaTecnica {
  id: string;
  produto_id: string;
  /** @nullable */
  produto_nome?: string | null;
  /** @nullable */
  rendimento?: number | null;
  /** @nullable */
  unidade_rendimento?: string | null;
  /** @nullable */
  cmv_total?: number | null;
  /** @nullable */
  created_at?: string | null;
}

export interface FichaItemDetalhado {
  id: string;
  insumo_id: string;
  /** @nullable */
  insumo_nome?: string | null;
  /** @nullable */
  unidade?: string | null;
  /** @nullable */
  preco_unitario?: number | null;
  /** @nullable */
  fator_correcao?: number | null;
  quantidade: number;
  /** @nullable */
  custo_item?: number | null;
}

export interface FichaTecnicaDetalhada {
  id: string;
  produto_id: string;
  /** @nullable */
  produto_nome?: string | null;
  /** @nullable */
  rendimento?: number | null;
  /** @nullable */
  unidade_rendimento?: string | null;
  /** @nullable */
  cmv_total?: number | null;
  itens: FichaItemDetalhado[];
  /** @nullable */
  created_at?: string | null;
}

export interface FichaInput {
  produto_id: string;
  /** @nullable */
  rendimento?: number | null;
  /** @nullable */
  unidade_rendimento?: string | null;
}

export interface FichaItem {
  id: string;
  ficha_id: string;
  insumo_id: string;
  quantidade: number;
}

export interface FichaItemInput {
  insumo_id: string;
  quantidade: number;
}

export interface Despesa {
  id: string;
  descricao: string;
  valor: number;
  /** @nullable */
  categoria?: string | null;
  /** @nullable */
  created_at?: string | null;
}

export interface DespesaInput {
  descricao: string;
  valor: number;
  /** @nullable */
  categoria?: string | null;
}

export type LancamentoTipo = typeof LancamentoTipo[keyof typeof LancamentoTipo];


export const LancamentoTipo = {
  receita: 'receita',
  despesa: 'despesa',
} as const;

export interface Lancamento {
  id: string;
  descricao: string;
  tipo: LancamentoTipo;
  valor: number;
  data: string;
  /** @nullable */
  created_at?: string | null;
}

export type LancamentoInputTipo = typeof LancamentoInputTipo[keyof typeof LancamentoInputTipo];


export const LancamentoInputTipo = {
  receita: 'receita',
  despesa: 'despesa',
} as const;

export interface LancamentoInput {
  descricao: string;
  tipo: LancamentoInputTipo;
  valor: number;
  data: string;
}

export interface DashboardSummary {
  receita_total: number;
  custos_totais: number;
  margem_media: number;
  resultado_mes: number;
  total_produtos: number;
  total_insumos: number;
  total_fichas: number;
}

export interface ProdutoRanking {
  id: string;
  nome: string;
  /** @nullable */
  categoria?: string | null;
  preco_venda: number;
  margem_contribuicao: number;
  margem_pct: number;
}

export type AlertaMargemNivel = typeof AlertaMargemNivel[keyof typeof AlertaMargemNivel];


export const AlertaMargemNivel = {
  critico: 'critico',
  atencao: 'atencao',
} as const;

export interface AlertaMargem {
  id: string;
  nome: string;
  margem_pct: number;
  nivel: AlertaMargemNivel;
}

export interface FluxoDiario {
  data: string;
  receita: number;
  despesa: number;
  resultado: number;
}

export interface PontoEquilibrio {
  despesas_fixas_total: number;
  margem_media: number;
  ponto_contabil: number;
  ponto_economico: number;
  unidades_necessarias: number;
}

export interface Assinatura {
  id: string;
  plano: string;
  status: string;
  /** @nullable */
  valido_ate?: string | null;
}