# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Precifica API - Financial management and recipe costing for small food businesses
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /insumos:
    get:
      operationId: listInsumos
      tags: [insumos]
      summary: List ingredients
      responses:
        "200":
          description: List of ingredients
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Insumo"
    post:
      operationId: createInsumo
      tags: [insumos]
      summary: Create an ingredient
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InsumoInput"
      responses:
        "201":
          description: Created ingredient
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Insumo"
  /insumos/{id}:
    put:
      operationId: updateInsumo
      tags: [insumos]
      summary: Update an ingredient
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InsumoInput"
      responses:
        "200":
          description: Updated ingredient
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Insumo"
    delete:
      operationId: deleteInsumo
      tags: [insumos]
      summary: Delete an ingredient
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
  /fichas:
    get:
      operationId: listFichas
      tags: [fichas]
      summary: List recipe sheets
      responses:
        "200":
          description: List of recipe sheets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FichaTecnica"
    post:
      operationId: createFicha
      tags: [fichas]
      summary: Create a recipe sheet
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FichaInput"
      responses:
        "201":
          description: Created recipe sheet
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FichaTecnica"
  /fichas/{id}:
    get:
      operationId: getFicha
      tags: [fichas]
      summary: Get recipe sheet with items
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Recipe sheet with items
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FichaTecnicaDetalhada"
    put:
      operationId: updateFicha
      tags: [fichas]
      summary: Update recipe sheet
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FichaInput"
      responses:
        "200":
          description: Updated recipe sheet
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FichaTecnica"
    delete:
      operationId: deleteFicha
      tags: [fichas]
      summary: Delete a recipe sheet
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
  /fichas/{id}/itens:
    post:
      operationId: addFichaItem
      tags: [fichas]
      summary: Add item to recipe sheet
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FichaItemInput"
      responses:
        "201":
          description: Created item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FichaItem"
  /fichas/{id}/itens/{itemId}:
    put:
      operationId: updateFichaItem
      tags: [fichas]
      summary: Update item in recipe sheet
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: itemId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FichaItemInput"
      responses:
        "200":
          description: Updated item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FichaItem"
    delete:
      operationId: deleteFichaItem
      tags: [fichas]
      summary: Delete item from recipe sheet
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: itemId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
  /despesas:
    get:
      operationId: listDespesas
      tags: [despesas]
      summary: List fixed costs
      responses:
        "200":
          description: List of fixed costs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Despesa"
    post:
      operationId: createDespesa
      tags: [despesas]
      summary: Create a fixed cost
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DespesaInput"
      responses:
        "201":
          description: Created fixed cost
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Despesa"
  /despesas/{id}:
    put:
      operationId: updateDespesa
      tags: [despesas]
      summary: Update a fixed cost
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DespesaInput"
      responses:
        "200":
          description: Updated fixed cost
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Despesa"
    delete:
      operationId: deleteDespesa
      tags: [despesas]
      summary: Delete a fixed cost
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    Insumo:
      type: object
      properties:
        id:
          type: string
        nome:
          type: string
        unidade:
          type: string
        preco_unitario:
          type: number
        fator_correcao:
          type: number
        created_at:
          type: ["string", "null"]
      required: [id, nome, unidade, preco_unitario, fator_correcao]
    InsumoInput:
      type: object
      properties:
        nome:
          type: string
        unidade:
          type: string
        preco_unitario:
          type: number
        fator_correcao:
          type: number
      required: [nome, unidade, preco_unitario, fator_correcao]
    FichaTecnica:
      type: object
      properties:
        id:
          type: string
        produto_id:
          type: string
        produto_nome:
          type: ["string", "null"]
        rendimento:
          type: ["number", "null"]
        unidade_rendimento:
          type: ["string", "null"]
        cmv_total:
          type: ["number", "null"]
        created_at:
          type: ["string", "null"]
      required: [id, produto_id]
    FichaTecnicaDetalhada:
      type: object
      properties:
        id:
          type: string
        produto_id:
          type: string
        produto_nome:
          type: ["string", "null"]
        rendimento:
          type: ["number", "null"]
        unidade_rendimento:
          type: ["string", "null"]
        cmv_total:
          type: ["number", "null"]
        itens:
          type: array
          items:
            $ref: "#/components/schemas/FichaItemDetalhado"
        created_at:
          type: ["string", "null"]
      required: [id, produto_id, itens]
    FichaInput:
      type: object
      properties:
        produto_id:
          type: string
        rendimento:
          type: ["number", "null"]
        unidade_rendimento:
          type: ["string", "null"]
      required: [produto_id]
    FichaItem:
      type: object
      properties:
        id:
          type: string
        ficha_id:
          type: string
        insumo_id:
          type: string
        quantidade:
          type: number
      required: [id, ficha_id, insumo_id, quantidade]
    FichaItemDetalhado:
      type: object
      properties:
        id:
          type: string
        insumo_id:
          type: string
        insumo_nome:
          type: ["string", "null"]
        unidade:
          type: ["string", "null"]
        preco_unitario:
          type: ["number", "null"]
        fator_correcao:
          type: ["number", "null"]
        quantidade:
          type: number
        custo_item:
          type: ["number", "null"]
      required: [id, insumo_id, quantidade]
    FichaItemInput:
      type: object
      properties:
        insumo_id:
          type: string
        quantidade:
          type: number
      required: [insumo_id, quantidade]
    Despesa:
      type: object
      properties:
        id:
          type: string
        descricao:
          type: string
        valor:
          type: number
        categoria:
          type: ["string", "null"]
        created_at:
          type: ["string", "null"]
      required: [id, descricao, valor]
    DespesaInput:
      type: object
      properties:
        descricao:
          type: string
        valor:
          type: number
        categoria:
          type: ["string", "null"]
      required: [descricao, valor]
...