# lib/api-zod/src/generated/types/despesa.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Precifica API - Financial management and recipe costing for small food businesses
 * OpenAPI spec version: 0.1.0
 */

export interface Despesa {
  id: string;
  descricao: string;
  valor: number;
  /** @nullable */
  categoria?: string | null;
  /** @nullable */
  created_at?: string | null;
}