# lib/api-zod/src/generated/types/fichaItem.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Precifica API - Financial management and recipe costing for small food businesses
 * OpenAPI spec version: 0.1.0
 */

export interface FichaItem {
  id: string;
  ficha_id: string;
  insumo_id: string;
  quantidade: number;
}