# lib/api-zod/src/generated/types/fichaItemDetalhado.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Precifica API - Financial management and recipe costing for small food businesses
 * OpenAPI spec version: 0.1.0
 */

export interface FichaItemDetalhado {
  id: string;
  insumo_id: string;
  /** @nullable */
  insumo_nome?: string | null;
  /** @nullable */
  unidade?: string | null;
  /** @nullable */
  preco_unitario?: number | null;
  /** @nullable */
  fator_correcao?: number | null;
  quantidade: number;
  /** @nullable */
  custo_item?: number | null;
}