# lib/api-zod/src/generated/types/insumoInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Precifica API - Financial management and recipe costing for small food businesses
 * OpenAPI spec version: 0.1.0
 */

export interface InsumoInput {
  nome: string;
  unidade: string;
  preco_unitario: number;
  fator_correcao: number;
}